Assign --fixed-offset to 1 (Jan 1, This year)
The old --fixed-offset value of 163, generates post-dated certificates
before the 12/13 of June each year. These certificates always fail to
verify because they are not valid at the time of validation. IE. Now.

This sets the fixed offset to January 1 of "this year", so that all
certificates are valid at the time of validation.

Also, remove the error exception when validating, all certificates
must be valid and the --fixed-offset date code is exercised.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>